fake-review-detector-api/pyproject.toml
[tool.poetry]
name = "fake_review_detector_api"
version = "0.1.0"
description = ""
authors = ["Ibrahim Haji Abdi <ibrahim.hajiabdi09@gmail.com>"]
packages = [{include = "fake_review_detector_api"}]

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "3.15.1"
packaging = "24.0"
pip-autoremove = "0.10.0"
requests = "2.31.0"
asgiref = "3.8.1"
django = "5.0.4"
djangorestframework-api-key = "3.0.0"
python-dotenv = "1.0.1"
sqlparse = "0.5.0"
typing-extensions = "4.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"